specify min Mixin::Linewise requirement
Needed for t/reader-err.t to pass.

The bom check added in 1ec26331 is for UTF-8 input, and since 226272d1
t/reader-err.t checks for this. But that test only passes if the input
actually is UTF-8. Mixin::Linewise changed to defaulting to UTF-8 in
0.100.

With older versions of Mixin::Linewise, the input stream wasn't UTF-8, so
the first character read wasn't a bom, but just the byte EF. So no ‘no
bom’ error was thrown, and the test failed.

Issue #7